1.基于aml的更新进行评测通用组件的优化 2.优化验证通用组件并保持一致性 3.删除冗余组件
